# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chemidot B2B Chemical Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: country
          in: query
          schema:
            type: string
        - name: minMoq
          in: query
          schema:
            type: number
        - name: maxMoq
          in: query
          schema:
            type: number
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: certifications
          in: query
          schema:
            type: string
        - name: collectiveEligible
          in: query
          schema:
            type: boolean
        - name: verifiedSupplier
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, moq_asc, rating]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (supplier only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects and case studies
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: industry
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project/case study by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    PricingTier:
      type: object
      properties:
        minQuantity:
          type: number
        maxQuantity:
          type: number
          nullable: true
        pricePerUnit:
          type: number
        discountPercent:
          type: number
      required: [minQuantity, pricePerUnit, discountPercent]
    TechnicalSpec:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        unit:
          type: string
          nullable: true
      required: [name, value]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        casNumber:
          type: string
          nullable: true
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        supplierVerified:
          type: boolean
        supplierCountry:
          type: string
        imageUrl:
          type: string
          nullable: true
        moq:
          type: number
        moqUnit:
          type: string
        basePrice:
          type: number
          nullable: true
        currency:
          type: string
        availability:
          type: string
          enum: [in_stock, limited, out_of_stock]
        deliveryLeadTime:
          type: string
        collectiveEligible:
          type: boolean
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, description, categoryId, categoryName, supplierId, supplierName, supplierVerified, supplierCountry, moq, moqUnit, currency, availability, deliveryLeadTime, collectiveEligible, reviewCount, createdAt]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
            technicalSpecs:
              type: array
              items:
                $ref: "#/components/schemas/TechnicalSpec"
            pricingTiers:
              type: array
              items:
                $ref: "#/components/schemas/PricingTier"
            applications:
              type: array
              items:
                type: string
            packaging:
              type: string
              nullable: true
            countryOfOrigin:
              type: string
            sdsDocumentUrl:
              type: string
              nullable: true
            supplier:
              $ref: "#/components/schemas/Supplier"
          required: [images, technicalSpecs, pricingTiers, applications, countryOfOrigin]
    ProductsPage:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        casNumber:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        moq:
          type: number
        moqUnit:
          type: string
        basePrice:
          type: number
        currency:
          type: string
        availability:
          type: string
          enum: [in_stock, limited, out_of_stock]
        deliveryLeadTime:
          type: string
        collectiveEligible:
          type: boolean
        countryOfOrigin:
          type: string
        technicalSpecs:
          type: array
          items:
            $ref: "#/components/schemas/TechnicalSpec"
        pricingTiers:
          type: array
          items:
            $ref: "#/components/schemas/PricingTier"
        applications:
          type: array
          items:
            type: string
        packaging:
          type: string
      required: [name, description, categoryId, moq, moqUnit, currency, availability, deliveryLeadTime, collectiveEligible, countryOfOrigin]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        moq:
          type: number
        basePrice:
          type: number
        availability:
          type: string
          enum: [in_stock, limited, out_of_stock]
        deliveryLeadTime:
          type: string
        collectiveEligible:
          type: boolean
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        logoUrl:
          type: string
          nullable: true
        country:
          type: string
        verified:
          type: boolean
        certifications:
          type: array
          items:
            type: string
        productCount:
          type: integer
        responseRate:
          type: number
        avgResponseTime:
          type: string
        yearsInBusiness:
          type: integer
          nullable: true
        rating:
          type: number
          nullable: true
      required: [id, companyName, country, verified, certifications, productCount, responseRate, avgResponseTime]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        buyerId:
          type: integer
        buyerCompanyName:
          type: string
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, productId, buyerId, buyerCompanyName, rating, createdAt]
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
      required: [rating]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        description:
          type: string
        heroImageUrl:
          type: string
          nullable: true
        industryTags:
          type: array
          items:
            type: string
        chemicalsUsed:
          type: array
          items:
            type: string
        supplierId:
          type: integer
          nullable: true
        featured:
          type: boolean
        createdAt:
          type: string
      required: [id, title, slug, summary, description, industryTags, chemicalsUsed, featured, createdAt]